# openapi.yaml
openapi: 3.0.3
info: {title: BYOD Server API, version: 0.6.0}
servers: [{url: https://exam.cs.ac.cn}]
security: [{userSession: []}]
paths:
...
  /auth/me:
    get: {security: [], operationId: getCurrentUser, responses: {'200': {description: Current user, content: {application/json: {schema: {$ref: '#/components/schemas/AuthMe'}}}}, '401': {description: Login required}}}
  /auth/logout:
    post: {security: [{userSession: []}], operationId: logoutUser, responses: {'204': {description: Logged out}}}
...
  /admin/api/users:
    get: {operationId: listUsers, parameters: [{name: q, in: query, required: false, schema: {type: string}}, {name: page, in: query, required: false, description: 1-based page number (default 1), schema: {type: integer, minimum: 1}}, {name: page_size, in: query, required: false, description: "Items per page (default 50, maximum 100)", schema: {type: integer, minimum: 1, maximum: 100}}, {name: limit, in: query, required: false, description: Legacy alias for page_size, schema: {type: integer, minimum: 1, maximum: 100}}, {name: offset, in: query, required: false, description: Legacy zero-based item offset, schema: {type: integer, minimum: 0}}], responses: {'200': {description: Users page; the JSON body remains an array for compatibility, and pagination metadata is returned in X-Total-Count, X-Page, X-Page-Size, and X-Has-More headers., headers: {X-Total-Count: {schema: {type: integer}}, X-Page: {schema: {type: integer}}, X-Page-Size: {schema: {type: integer}}, X-Has-More: {schema: {type: boolean}}}, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/User'}}}}}}}
    post: {operationId: inviteUser, requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserInput'}}}}, responses: {'201': {description: Invited user, content: {application/json: {schema: {$ref: '#/components/schemas/User'}}}}}}
  /admin/api/users/{userId}:
    parameters: [{name: userId, in: path, required: true, schema: {type: string}}]
    get: {operationId: getUser, responses: {'200': {description: User, content: {application/json: {schema: {$ref: '#/components/schemas/User'}}}}}}
    patch: {operationId: updateUser, requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UserUpdate'}}}}, responses: {'200': {description: Updated user, content: {application/json: {schema: {$ref: '#/components/schemas/User'}}}}}}
...
components:
  securitySchemes: {adminToken: {type: apiKey, in: header, name: X-Admin-Token}, userSession: {type: apiKey, in: cookie, name: __Host-byod_user}, browserToken: {type: http, scheme: bearer}}
  schemas:
...
    User: {type: object, required: [id, issuer, role, platform_admin, enabled, display_name, nickname, picture, created_at], properties: {id: {type: string}, issuer: {type: string}, subject: {type: string, nullable: true}, email: {type: string, format: email, nullable: true}, display_name: {type: string}, nickname: {type: string, description: OIDC nickname claim}, picture: {type: string, description: OIDC picture claim}, role: {type: string, deprecated: true, description: Compatibility projection; use platform_admin and exam-admin memberships., enum: [student, admin]}, platform_admin: {type: boolean}, enabled: {type: boolean}, created_at: {type: string, format: date-time}, last_login_at: {type: string, format: date-time, nullable: true}}}
    UserInput: {type: object, required: [email], properties: {email: {type: string, format: email}, display_name: {type: string}, platform_admin: {type: boolean}, role: {type: string, deprecated: true, enum: [student, admin]}}}
    UserUpdate: {type: object, properties: {display_name: {type: string}, platform_admin: {type: boolean}, role: {type: string, deprecated: true, enum: [student, admin]}, enabled: {type: boolean}}}
...
    AuthMe: {type: object, required: [user, csrf_token, capabilities], properties: {user: {$ref: '#/components/schemas/User'}, csrf_token: {type: string}, capabilities: {type: object, required: [platform_admin, exam_admin], properties: {platform_admin: {type: boolean}, exam_admin: {type: boolean}}}}}
...